move upload view tests to test_routes
